+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `selli/ui` will be documented in this file.
+
+## Unreleased
+
+- Initial package foundation: Selli Design System CSS tokens (colors, typography, spacing,
+  effects, density, base), service provider, config, facade and icon set.